fix: 멀티모듈 validator 에러 해결
