+++ bbiyong-api/build.gradle
@@ -11,9 +11,8 @@ repositories {
 
 dependencies {
     implementation project(':bbiyong-domain')
+    implementation 'org.springdoc:springdoc-openapi-starter-webmvc-ui:2.0.2'
     implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
-    implementation 'org.springdoc:springdoc-openapi-ui:1.6.12'
-    runtimeOnly 'com.mysql:mysql-connector-j'
 }
 
 test {
