+++ bbiyong-api/src/main/java/com/capstone/common/config/SwaggerConfig.java
@@ -1,27 +1,28 @@
 package com.capstone.common.config;
 
-import io.swagger.v3.oas.models.Components;
 import io.swagger.v3.oas.models.OpenAPI;
 import io.swagger.v3.oas.models.info.Info;
+import org.springdoc.core.models.GroupedOpenApi;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
 @Configuration
-public class SwaggerConfig implements WebMvcConfigurer {
-
-    private static final String API_TITLE = "삐용삐용 API";
-    private static final String API_DESCRIPTION = "삐용삐용 API 문서";
-    private static final String API_VERSION = "Version 0.0.1";
+public class SwaggerConfig {
 
     @Bean
-    public OpenAPI customOpenAPI() {
+    public GroupedOpenApi publicApi() {
+        return GroupedOpenApi.builder()
+                .group("v1")
+                .pathsToMatch("/**")
+                .build();
+    }
+
+    @Bean
+    public OpenAPI springOpenAPI() {
         return new OpenAPI()
-                .components(new Components())
-                .info(new Info()
-                        .title(API_TITLE)
-                        .description(API_DESCRIPTION)
-                        .version(API_VERSION));
+                .info(new Info().title("삐용삐용 API")
+                        .description("삐용삐용 API 명세서입니다.")
+                        .version("v0.0.1"));
     }
 
 }
