--- bbiyong-api/src/main/resources/application.yml
@@ -1,12 +1,0 @@
-spring:
-  jpa:
-    show-sql: false
-    database-platform: org.hibernate.dialect.MySQL5InnoDBDialect
-    open-in-view: false
-    hibernate:
-      ddl-auto: update
-    properties:
-      hibernate:
-        dialect: org.hibernate.dialect.MySQL5InnoDBDialect
-  config:
-    import: application-datasource.yml
