+++ bbiyong-domain/build.gradle
@@ -7,5 +7,5 @@ repositories {
 
 dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
-    implementation 'org.hibernate:hibernate-core:6.2.2.Final'
+    runtimeOnly 'com.mysql:mysql-connector-j'
 }
