+++ bbiyong-domain/src/main/java/com/capstone/emerMsg/domain/EmerMsg.java
@@ -1,42 +1,47 @@
-package com.capstone.metro;
+package com.capstone.emerMsg.domain;
 
 import com.capstone.common.DateTimeEntity;
+import com.capstone.location.domain.Location;
 import jakarta.persistence.*;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-import java.time.LocalDateTime;
+import java.util.Date;
 
 import static jakarta.persistence.GenerationType.IDENTITY;
 
+@Entity
 @Getter
 @NoArgsConstructor
-@Entity
-public class Metro extends DateTimeEntity {
+public class EmerMsg extends DateTimeEntity {
     @Id
     @GeneratedValue(strategy = IDENTITY)
-    @Column(nullable = false)
+    @Column(name = "emer_msg_id", nullable = false)
     private Long id;
 
     @Column(nullable = false)
     private Long openapiId;
 
     @Column(nullable = false)
-    private String text;
+    private String message;
 
     @Column(nullable = false)
-    private LocalDateTime startDate;
+    private Date startDate;
 
     @Column(nullable = false)
-    private LocalDateTime endDate;
+    private Date endDate;
 
+    @ManyToOne
+    @JoinColumn(name = "location_id")
+    private Location location;
 
     @Builder
-    public Metro(Long openapiId, String text, LocalDateTime startDate, LocalDateTime endDate) {
+    public EmerMsg(Long openapiId, String message, Date startDate, Date endDate, Location location) {
         this.openapiId = openapiId;
-        this.text = text;
+        this.message = message;
         this.startDate = startDate;
         this.endDate = endDate;
+        this.location = location;
     }
 }
