+++ bbiyong-domain/src/main/java/com/capstone/location/repository/LocationRepository.java
@@ -1,6 +1,6 @@
 package com.capstone.location.repository;
 
-import com.capstone.location.Location;
+import com.capstone.location.domain.Location;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface LocationRepository extends JpaRepository<Location, Long> {
