+++ build.gradle
@@ -38,6 +38,7 @@ subprojects {
 		compileOnly 'org.projectlombok:lombok'
 		annotationProcessor 'org.projectlombok:lombok'
 		testImplementation 'junit:junit:4.13.1'
+		implementation 'org.hibernate.validator:hibernate-validator:7.0.1.Final'
 	}
 }
